update 180454 to 7165

+++ bio/sources/pubmed-gene/main/src/org/intermine/bio/dataconversion/PubMedGeneConverter.java
@@ -84,7 +84,7 @@ public class PubMedGeneConverter extends FileConverter
         // D.melanogaster
         organismsToProcess.add(new Integer(7227));
         // A.gambiae str PEST
-        organismsToProcess.add(new Integer(9165));
+        organismsToProcess.add(new Integer(7165));
 
         Item datasource = createItem("DataSource");
         datasource.setAttribute("name", "NCBI");
